requirements.txt
# SIM-KB Server Requirements
flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
rich>=13.0.0
